requirements.txt
asgiref==3.3.1
Django==3.1.7
django-countries==7.1
django-crispy-forms==1.11.2
django-js-asset==1.2.2
django-mptt==0.12.0
django-widget-tweaks==1.4.8
pep8==1.7.1
Pillow==8.2.0
psycopg2==2.8.6
pycodestyle==2.7.0
pytz==2021.1
sqlparse==0.4.1
